+++ lib/gooddata/bricks/middleware/fs_upload_middleware.rb
@@ -8,7 +8,7 @@ require_relative 'base_middleware'
 
 module GoodData
   module Bricks
-    class FsUploadMiddleware < Bricks::Middleware
+    class FsProjectUploadMiddleware < Bricks::Middleware
       def initialize(options = {})
         super
         @destination = options[:destination]
@@ -18,16 +18,17 @@ module GoodData
         returning(@app.call(params)) do |_result|
           destination = @destination
           (params['gdc_files_to_upload'] || []).each do |f|
-            path = f[:path]
             case destination.to_s
             when 'staging'
-              url = GoodData.get_user_webdav_path(path)
-              GoodData.upload_to_user_webdav(path)
-              puts "Uploaded local file \"#{path}\" to url \"#{url + path}\""
+              GoodData.upload_to_project_webdav(f[:path], directory: f[:webdav_directory])
+              puts "Uploaded local file \"#{f[:path]}\" to webdav."
             end
           end
         end
       end
     end
+
+    # Alias to make it backwards compatible
+    FsUploadMiddleware = FsProjectUploadMiddleware
   end
 end